Reverting to original modal implementation
